Move mgxllama2 components to multiple files
